+++ httprunner/cli.py
@@ -6,10 +6,10 @@ import pytest
 from loguru import logger
 
 from httprunner import __description__, __version__
-from httprunner.ext.har2case import init_har2case_parser, main_har2case
-from httprunner.ext.make import init_make_parser, main_make
-from httprunner.ext.scaffold import init_parser_scaffold, main_scaffold
 from httprunner.compat import ensure_cli_args
+from httprunner.ext.har2case import init_har2case_parser, main_har2case
+from httprunner.ext.scaffold import init_parser_scaffold, main_scaffold
+from httprunner.make import init_make_parser, main_make
 
 
 def init_parser_run(subparsers):
